Make dashboard elevator display stuff a double

+++ src/com/_604robotics/robot2018/modules/Dashboard.java
@@ -25,8 +25,8 @@ public class Dashboard extends DashboardModule {
     public final Input<Double> elevatorDownwardsRange = addDashboardInput("Downwards Range", 0.0);
     public final Input<Boolean> elevatorFailsafed = addDashboardInput("Failsafed", false);
     
-    public final Input<Integer> elevatorRate = addDashboardInput("Elevator Rate", 0);
-    public final Input<Integer> elevatorClicks = addDashboardInput("Elevator Clicks", 0);
+    public final Input<Double> elevatorRate = addDashboardInput("Elevator Rate", 0.0);
+    public final Input<Double> elevatorClicks = addDashboardInput("Elevator Clicks", 0.0);
     
     public final Input<Boolean> holding = addDashboardInput("Holding", true);
     public final Input<Double> power = addDashboardInput("Elevator Power", 0.0);
